ДАС Домашно 4/requirements.txt
flask
pandas
numpy
beautifulsoup4
selenium
python-dateutil
requests
ta
gunicorn
python-dotenv